# internal/repository/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.17.2

package sqlc

import (
	"time"

	null "gopkg.in/guregu/null.v4"
)

type Url struct {
	ID             int32       `json:"id"`
	ShortUri       null.String `json:"short_uri"`
	UserID         int32       `json:"user_id"`
	RequestedCount int32       `json:"requested_count"`
	OriginalUrl    string      `json:"original_url"`
	ExpiresAt      time.Time   `json:"expires_at"`
}

type User struct {
	ID       int32  `json:"id"`
	Username string `json:"username"`
	FullName string `json:"full_name"`
	Password string `json:"password"`
}